==> methodology_blank/update_existing/1.1_Downloading_tweets_for_new_period.ipynb <==
import json
from twarc_csv import DataFrameConverter
from twarc.expansions import ensure_flattened
import pandas as pd
from twarc_csv import dataframe_converter
from twarc_csv import CSVConverter, DataFrameConverter
import math
import emoji
from operator import itemgetter
import matplotlib.pyplot as plt
from datetime import datetime
from datetime import timedelta
from advertools import extract_emoji
import numpy as np
import pickle

**Manual Input**: Set the path.

path='C:/Users/acoub/OneDrive/Desktop/DSDM/Thesis/ChileGov/methodology_blank'
sep='/'

path_scripts = path+sep+'update_existing'
path_data= path+sep+'data'

**Manual Input**: Set the start and end date for the upload.

start_date="YYYY-MM-DD"
end_date="YYYY-MM-DD"

filename = path_data+sep+'start_update_date.sav'
pickle.dump((start_date), open(filename, 'wb'))

filename = path_data+sep+'end_update_date.sav'
pickle.dump((end_date), open(filename, 'wb'))

file = path_data+sep+'keywords.sav'
list_of_words=pickle.load(open(file, 'rb'))

file = path_data+sep+'country_string.sav'
str_country=pickle.load(open(file, 'rb'))

query=""
for word in list_of_words:
    query+= word + " OR "
query=query[:len(query)-4]

query_place="("+query+")"+ " -is:retweet place_country:CL"
query_wrelated="("+query+")"+ str_country + " -is:retweet"

with open(path_data+sep+"search_1_updating.txt", 'w') as f:
    f.write(query_place)
    f.write("\n")
    f.write(query_wrelated)

**Manual revision:** Run the following two cells to know the total number of tweets to download and see if does not exceed the limit of your twarc acount. If is OK run also the next cell.

file=path_data+sep+"search_1_updating.txt"
%cd {path_data}
! twarc2 searches --archive --counts-only --granularity day --start-time {start_date} --end-time {end_date} {file} search_updating_count.csv

count=pd.read_csv(path_data+sep+'search_updating_count.csv',encoding = "ISO-8859-1")
print("Tweets to download: "+str(count['day_count'].sum()))

**Download the tweets:** (Only if you have enought limit in your twarc account)

%cd {path_data}
! twarc2 searches --archive --start-time {start_date} --end-time {end_date} {file} search_1_uploading.json
